code cleanup, compile fixes

+++ Hazelcast.Net/Hazelcast.Client.Request.Concurrent.Lock/UnlockRequest.cs
@@ -5,11 +5,13 @@ namespace Hazelcast.Client.Request.Concurrent.Lock
 {
     internal sealed class UnlockRequest : AbstractUnlockRequest
     {
-        public UnlockRequest(IData key, int threadId) : base(key, threadId)
+        public UnlockRequest(IData key, long threadId)
+            : base(key, threadId)
         {
         }
 
-        public UnlockRequest(IData key, int threadId, bool force) : base(key, threadId, force)
+        public UnlockRequest(IData key, long threadId, bool force)
+            : base(key, threadId, force)
         {
         }
